revision about type of struc parameters

+++ type/auth.d.ts
@@ -17,15 +17,15 @@ export class CertificateEntry extends Struct {
     /**
      * new CertificateEntry
      * @param {Uint8Array} certificate - type either "ASN1 SPKI" or "X509"
-     * @param {Uint16} extensions - optional with default value of Uint8Array([0,0])
+     * @param {Uint8Array} extensions - optional with default value of Uint8Array([0,0])
      */
-    static a(certificate: Uint8Array, extensions?: Uint16): CertificateEntry;
+    static a(certificate: Uint8Array, extensions?: Uint8Array): CertificateEntry;
     /**
      *
      * @param {Uint8Array} certificate - type either "ASN1 SPKI" or "X509"
-     * @param {Uint16} extensions - optional with default value of Uint8Array([0,0])
+     * @param {Uint8Array} extensions - optional with default value of Uint8Array([0,0])
      */
-    constructor(certificate: Uint8Array, extensions?: Uint16);
+    constructor(certificate: Uint8Array, extensions?: Uint8Array);
 }
 /**
  * Certificate
